HMSW-4286 : fix data type dismatch.

+++ tools/llm_perf/include/llm/HmllmInferMultiBatch.h
@@ -103,7 +103,7 @@ class HmllmInferMultiBatch : public HmllmInferBase {
   // Configuration parameters - Model properties
   int prefill_length = 0;
   int embedding_length = 0;
-  int context_max_length = 0;
+  long int context_max_length = 0;
   // Batch size
   int batch = 0;
   int argmax_dim_len = 0;
